Delete comfirmation with sweetalert2

+++ src/app/products/products.component.ts
@@ -4,7 +4,7 @@ import { Product } from '../product';
 import { Observable } from 'rxjs';
 import { Router } from '@angular/router';
 import { ProductDetailComponent } from '../product-detail/product-detail.component';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-products',
   templateUrl: './products.component.html',
@@ -26,13 +26,38 @@ export class ProductsComponent implements OnInit {
   }
 
   deleteProduct(id: number) {
-    this.apiService.deleteProduct(id)
+    Swal.fire({
+      title: 'Delete Confirmation',
+      text: 'Comfirm to delete data?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire(
+          'Deleted!',
+          'Product data has been deleted.',
+          'success'
+        );
+        this.apiService.deleteProduct(id)
+          .subscribe(
+          data => {
+            console.log(data);
+            this.reloadData();
+          },
+          error => console.log(error));
+      }
+    });
+    // default
+    /*this.apiService.deleteProduct(id)
       .subscribe(
         data => {
           console.log(data);
           this.reloadData();
         },
-        error => console.log(error));
+        error => console.log(error));*/
   }
 
   productDetail(id: number) {
